refactor: extract IdentityAttributeValueValidator
