+++ packages/runtime/src/useCases/consumption/attributes/CreateRepositoryAttribute.ts
@@ -1,12 +1,13 @@
-import { ApplicationError, Result } from "@js-soft/ts-utils";
+import { Result } from "@js-soft/ts-utils";
 import { AttributesController, CreateRepositoryAttributeParams } from "@nmshd/consumption";
 import { AttributeValues } from "@nmshd/content";
 import { AccountController } from "@nmshd/transport";
 import { Inject } from "@nmshd/typescript-ioc";
 import { LocalAttributeDTO } from "../../../types";
-import { ISO8601DateTimeString, RuntimeErrors, SchemaRepository, SchemaValidator, UseCase, ValidationFailure, ValidationResult } from "../../common";
+import { ISO8601DateTimeString, RuntimeErrors, SchemaRepository, SchemaValidator, UseCase, ValidationResult } from "../../common";
 import { IValidator } from "../../common/validation/IValidator";
 import { AttributeMapper } from "./AttributeMapper";
+import { IdentityAttributeValueValidator } from "./IdentityAttributeValueValidator";
 
 interface AbstractCreateRepositoryAttributeRequest<T> {
     content: {
@@ -24,49 +25,13 @@ export interface SchemaValidatableCreateRepositoryAttributeRequest extends Abstr
 class Validator implements IValidator<CreateRepositoryAttributeRequest> {
     public constructor(@Inject private readonly schemaRepository: SchemaRepository) {}
 
-    public validate(value: CreateRepositoryAttributeRequest): Promise<ValidationResult> | ValidationResult {
+    public validate(request: CreateRepositoryAttributeRequest): Promise<ValidationResult> | ValidationResult {
         const requestSchemaValidator = new SchemaValidator(this.schemaRepository.getSchema("CreateRepositoryAttributeRequest"));
-        const requestValidationResult = requestSchemaValidator.validate(value);
+        const requestValidationResult = requestSchemaValidator.validate(request);
         if (requestValidationResult.isInvalid()) return requestValidationResult;
 
         const identityAttributeValueValidator = new IdentityAttributeValueValidator(this.schemaRepository);
-        return identityAttributeValueValidator.validate(value);
-    }
-}
-
-export class IdentityAttributeValueValidator implements IValidator<CreateRepositoryAttributeRequest> {
-    public constructor(@Inject private readonly schemaRepository: SchemaRepository) {}
-
-    public validate(value: CreateRepositoryAttributeRequest): Promise<ValidationResult> | ValidationResult {
-        const attributeValueType = value.content.value["@type"];
-        if (!AttributeValues.Identity.TYPE_NAMES.includes(attributeValueType)) {
-            const attributeValueTypeValidationResult = new ValidationResult();
-
-            attributeValueTypeValidationResult.addFailure(
-                new ValidationFailure(
-                    RuntimeErrors.general.invalidPropertyValue("content.value.@type must match one of the allowed Attribute value types for IdentityAttributes"),
-                    "@type"
-                )
-            );
-            return attributeValueTypeValidationResult;
-        }
-
-        const attributeValueSchemaValidator = new SchemaValidator(this.schemaRepository.getSchema(attributeValueType));
-        const attributeValueValidationResult = attributeValueSchemaValidator.validate(value.content.value);
-        return IdentityAttributeValueValidator.addPrefixToErrorMessagesOfResult(`${attributeValueType} :: `, attributeValueValidationResult);
-    }
-
-    private static addPrefixToErrorMessagesOfResult(prefix: string, validationResult: ValidationResult): ValidationResult {
-        if (validationResult.isValid()) return validationResult;
-
-        const failures = validationResult.getFailures();
-        const failuresWithPrefix = failures.map(
-            (failure) => new ValidationFailure(new ApplicationError(failure.error.code, `${prefix}${failure.error.message}`, failure.error.data), failure.propertyName)
-        );
-
-        const validationResultWithPrefix = new ValidationResult();
-        validationResultWithPrefix.addFailures(failuresWithPrefix);
-        return validationResultWithPrefix;
+        return identityAttributeValueValidator.validate(request.content.value);
     }
 }
